LayoutTests/fast/replaced/table-percent-height-expected.txt







Button
Button


















This is a searchable index. Enter search keywords: 
This is a searchable index. Enter search keywords: 




This test checks that replaced elements with percentage heights within table cells have the correct height.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyleForElement(document.getElementById('canvas-75'), 'width') is '300px'
PASS getComputedStyleForElement(document.getElementById('canvas-75'), 'height') is '112px'
PASS getComputedStyleForElement(document.getElementById('canvas-100'), 'width') is '300px'
PASS getComputedStyleForElement(document.getElementById('canvas-100'), 'height') is '150px'
PASS getComputedStyleForElement(document.getElementById('embed-75'), 'width') is '300px'
PASS getComputedStyleForElement(document.getElementById('embed-75'), 'height') is '112px'
PASS getComputedStyleForElement(document.getElementById('embed-100'), 'width') is '300px'
PASS getComputedStyleForElement(document.getElementById('embed-100'), 'height') is '150px'
PASS getComputedStyleForElement(document.getElementById('img-75'), 'width') is '75px'
PASS getComputedStyleForElement(document.getElementById('img-75'), 'height') is '75px'
PASS getComputedStyleForElement(document.getElementById('img-100'), 'width') is '100px'
PASS getComputedStyleForElement(document.getElementById('img-100'), 'height') is '100px'
PASS getComputedStyleForElement(document.getElementById('object-75'), 'width') is '300px'
PASS getComputedStyleForElement(document.getElementById('object-75'), 'height') is '112px'
PASS getComputedStyleForElement(document.getElementById('object-100'), 'width') is '300px'
PASS getComputedStyleForElement(document.getElementById('object-100'), 'height') is '150px'
PASS getComputedStyleForElement(document.getElementById('button-75'), 'width') is '51px'
PASS getComputedStyleForElement(document.getElementById('button-75'), 'height') is '15px'
PASS getComputedStyleForElement(document.getElementById('button-100'), 'width') is '51px'
PASS getComputedStyleForElement(document.getElementById('button-100'), 'height') is '15px'
PASS getComputedStyleForElement(document.getElementById('input-button-75'), 'width') is '16px'
PASS getComputedStyleForElement(document.getElementById('input-button-75'), 'height') is '18px'
PASS getComputedStyleForElement(document.getElementById('input-button-100'), 'width') is '16px'
PASS getComputedStyleForElement(document.getElementById('input-button-100'), 'height') is '18px'
PASS getComputedStyleForElement(document.getElementById('input-checkbox-75'), 'width') is '12px'
PASS getComputedStyleForElement(document.getElementById('input-checkbox-75'), 'height') is '4px'
PASS getComputedStyleForElement(document.getElementById('input-checkbox-100'), 'width') is '12px'
PASS getComputedStyleForElement(document.getElementById('input-checkbox-100'), 'height') is '6px'
PASS getComputedStyleForElement(document.getElementById('input-file-75'), 'width') is '237px'
PASS getComputedStyleForElement(document.getElementById('input-file-75'), 'height') is '13px'
PASS getComputedStyleForElement(document.getElementById('input-file-100'), 'width') is '237px'
PASS getComputedStyleForElement(document.getElementById('input-file-100'), 'height') is '18px'
PASS getComputedStyleForElement(document.getElementById('input-image-75'), 'width') is '75px'
PASS getComputedStyleForElement(document.getElementById('input-image-75'), 'height') is '75px'
PASS getComputedStyleForElement(document.getElementById('input-image-100'), 'width') is '100px'
PASS getComputedStyleForElement(document.getElementById('input-image-100'), 'height') is '100px'
PASS getComputedStyleForElement(document.getElementById('input-password-75'), 'width') is '159px'
PASS getComputedStyleForElement(document.getElementById('input-password-75'), 'height') is '8px'
PASS getComputedStyleForElement(document.getElementById('input-password-100'), 'width') is '159px'
PASS getComputedStyleForElement(document.getElementById('input-password-100'), 'height') is '13px'
PASS getComputedStyleForElement(document.getElementById('input-radio-75'), 'width') is '12px'
PASS getComputedStyleForElement(document.getElementById('input-radio-75'), 'height') is '4px'
PASS getComputedStyleForElement(document.getElementById('input-radio-100'), 'width') is '12px'
PASS getComputedStyleForElement(document.getElementById('input-radio-100'), 'height') is '6px'
PASS getComputedStyleForElement(document.getElementById('input-reset-75'), 'width') is '45px'
PASS getComputedStyleForElement(document.getElementById('input-reset-75'), 'height') is '18px'
PASS getComputedStyleForElement(document.getElementById('input-reset-100'), 'width') is '45px'
PASS getComputedStyleForElement(document.getElementById('input-reset-100'), 'height') is '18px'
PASS getComputedStyleForElement(document.getElementById('input-submit-75'), 'width') is '54px'
PASS getComputedStyleForElement(document.getElementById('input-submit-75'), 'height') is '18px'
PASS getComputedStyleForElement(document.getElementById('input-submit-100'), 'width') is '54px'
PASS getComputedStyleForElement(document.getElementById('input-submit-100'), 'height') is '18px'
PASS getComputedStyleForElement(document.getElementById('input-text-75'), 'width') is '159px'
PASS getComputedStyleForElement(document.getElementById('input-text-75'), 'height') is '8px'
PASS getComputedStyleForElement(document.getElementById('input-text-100'), 'width') is '159px'
PASS getComputedStyleForElement(document.getElementById('input-text-100'), 'height') is '13px'
PASS getComputedStyleForElement(document.getElementById('isindex-75'), 'width') is '153px'
PASS getComputedStyleForElement(document.getElementById('isindex-75'), 'height') is '35px'
PASS getComputedStyleForElement(document.getElementById('isindex-100'), 'width') is '153px'
PASS getComputedStyleForElement(document.getElementById('isindex-100'), 'height') is '49px'
PASS getComputedStyleForElement(document.getElementById('select-75'), 'width') is '68px'
PASS getComputedStyleForElement(document.getElementById('select-75'), 'height') is '18px'
PASS getComputedStyleForElement(document.getElementById('select-100'), 'width') is '68px'
PASS getComputedStyleForElement(document.getElementById('select-100'), 'height') is '18px'
PASS getComputedStyleForElement(document.getElementById('textarea-75'), 'width') is '161px'
PASS getComputedStyleForElement(document.getElementById('textarea-75'), 'height') is '19px'
PASS getComputedStyleForElement(document.getElementById('textarea-100'), 'width') is '161px'
PASS getComputedStyleForElement(document.getElementById('textarea-100'), 'height') is '28px'
PASS successfullyParsed is true

TEST COMPLETE
